Clear xgboost arch migrators status again
The macOS ARM migrator is already applied. However the bot shows an
error. We tried to clear this before, but the bot still ran into issues
with the recipe format. Have made a formatting change, which may help
(however was unable to reproduce the bot issue so couldn't test).
Clearing the status here to try again and see if it is fixed.